Fix CI

+++ odb-compiler/include/odb-compiler/parsers/db/Scanner.hpp
@@ -25,7 +25,7 @@ int dblex_init(dbscan_t* ptr_yy_globals);
 int dblex_init_extra(odb::db::Driver* db_user_defined, dbscan_t* ptr_yy_globals);
 int dblex_destroy(dbscan_t yyscanner);
 void dbset_in(FILE* _in_str, dbscan_t dbscanner);
-YY_BUFFER_STATE db_scan_bytes(const char *bytes, size_t len, dbscan_t dbscanner);
+YY_BUFFER_STATE db_scan_bytes(const char *bytes, int len, dbscan_t dbscanner);
 void db_delete_buffer(YY_BUFFER_STATE b , dbscan_t dbscanner);
 int dblex(DBSTYPE* dblval_param, DBLTYPE* yylloc_param, dbscan_t dbscanner);
 odb::db::Driver* dbget_extra(dbscan_t dbscanner);
